% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCAuth.R
\name{doRCAuth}
\alias{doRCAuth}
\title{Auth function for RECCap}
\usage{
doRCAuth(req, res, config)
}
\arguments{
\item{req}{Plumber req object}

\item{res}{Plumber res object}

\item{config}{\link{Config} object}
}
\value{
NULL if authenticated, a string URL to the REDCap project providing the authentication active link (see \link{RCAuthProcess}) 
for more information about the format of the REDCap active link).  Otherwise will return \code{list(err = "error message here")}
}
\description{
This function checks to see if the current session is authenticated, and if not, redirect them to REDCap to log in.  This can serve as a template for other authenication mechanisms
}
