% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCPID.R
\name{getPID}
\alias{getPID}
\title{Lookup PIDs in REDCap}
\usage{
getPID(sampleid, config)
}
\arguments{
\item{sampleid}{vector of length one or more of sample ids for which pids/mrns are needed}

\item{config}{Scrapert config}
}
\value{
tibble of columns sample_id and pid
}
\description{
Lookup PIDs in REDCap
}
