% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keyed_value}
\alias{keyed_value}
\alias{keyed_ts}
\title{Get Xpert Keyed Value}
\usage{
keyed_value(key, txt)

keyed_ts(key, txt, tz)
}
\arguments{
\item{key}{key of value to extract}

\item{txt}{source of extraction}
}
\value{
the value after the key
}
\description{
Get Xpert Keyed Value
}
\details{
\code{keyed_ts} returns the \code{POSIXct} date/time of the keyed value in UCT/GMT time
}
