% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Launch Scrapert}
\usage{
launch(wd = getwd(), server = FALSE)
}
\arguments{
\item{wd}{a working directory or not supplied to use the current working directory}

\item{server}{launch in server (daemon mode) `TRUE/FALSE`}
}
\value{
0 (zero) if successful
}
\description{
This function will launch Scrapert and log results to "scrapert.log" including installing any missing packages
}
\details{
The database, the config file, and the log will all be located in the working directory
}
\examples{
# Rscript -e 'Scrapert::launch()' 

# Scrapert::launch("/path/to/working_dir")

}
