% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modifiedXpert}
\alias{modifiedXpert}
\title{Add Optimized Xpert Results}
\usage{
modifiedXpert(df, config)
}
\arguments{
\item{df}{dataframe with Xpert data}

\item{config}{\link{Config} R6 class}
}
\value{
dataframe with \code{restricted_result} and \code{mod_ct_result} columns added
}
\description{
Add modificed Xpert results  to df using values in \link{Config}
}
