% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pidtest}
\alias{pidtest}
\title{PID Test Dataset}
\format{
## `pidtest`
A data frame with 4 rows and two columns
\describe{
  \item{sample_id}{Sample ID}
  \item{pid}{PID}
}
}
\usage{
pidtest
}
\description{
Small dataset of pids <--> sample ids to map with the included in Test_Result.pdf
}
\keyword{datasets}
