% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\name{RC}
\alias{RC}
\title{Scrapert REDCap R6 Class}
\description{
methods for interacting with REDCap
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RC-new}{\code{RC$new()}}
\item \href{#method-RC-hasConf}{\code{RC$hasConf()}}
\item \href{#method-RC-setKey}{\code{RC$setKey()}}
\item \href{#method-RC-auth}{\code{RC$auth()}}
\item \href{#method-RC-checkAuth}{\code{RC$checkAuth()}}
\item \href{#method-RC-getUser}{\code{RC$getUser()}}
\item \href{#method-RC-authfilter}{\code{RC$authfilter()}}
\item \href{#method-RC-getPID}{\code{RC$getPID()}}
\item \href{#method-RC-updateCRF}{\code{RC$updateCRF()}}
\item \href{#method-RC-CRFDelete}{\code{RC$CRFDelete()}}
\item \href{#method-RC-getPDF}{\code{RC$getPDF()}}
\item \href{#method-RC-backup}{\code{RC$backup()}}
\item \href{#method-RC-restore}{\code{RC$restore()}}
\item \href{#method-RC-clone}{\code{RC$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RC-new"></a>}}
\if{latex}{\out{\hypertarget{method-RC-new}{}}}
\subsection{Method \code{new()}}{
New RC class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RC$new(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{\link{Config} R6 object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
RC class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RC-hasConf"></a>}}
\if{latex}{\out{\hypertarget{method-RC-hasConf}{}}}
\subsection{Method \code{hasConf()}}{
Do we have enough setting to communicate with REDCap
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RC$hasConf(throw_if_not_set = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{throw_if_not_set}}{\code{logical} specifying if should return \code{FALSE} or throw errors}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if REDCap config set
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RC-setKey"></a>}}
\if{latex}{\out{\hypertarget{method-RC-setKey}{}}}
\subsection{Method \code{setKey()}}{
Set a new REDCap Project API Token
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RC$setKey(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{REDCAp project API token}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RC-auth"></a>}}
\if{latex}{\out{\hypertarget{method-RC-auth}{}}}
\subsection{Method \code{auth()}}{
check if REDCAp config is present (redirects otherwise to \code{RCConfig.html}).  if session cookie is set,
check to see if still authenticated every two minutes.  If not authenticated, will redirect to REDCap to reauthenicate
via the active linke
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RC$auth(req, res)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{Plumber req object}

\item{\code{res}}{Plumber res object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} if authenticated, other wise redirect to \code{RCConfig.html} or REDCap
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RC-checkAuth"></a>}}
\if{latex}{\out{\hypertarget{method-RC-checkAuth}{}}}
\subsection{Method \code{checkAuth()}}{
Process the REDCap active link.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RC$checkAuth(req, res)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{Plumber req object}

\item{\code{res}}{Plumber res object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The active link should post it's response to \code{http://localhost:port/authep}.  No other variables are required.
This will also set RCApi if it's posted as an request arg
}

\subsection{Returns}{
Will redirect if  api is posted as an arg after saving the config
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RC-getUser"></a>}}
\if{latex}{\out{\hypertarget{method-RC-getUser}{}}}
\subsection{Method \code{getUser()}}{
If able to get full name from REDCap will set session cookie along with username, 
othersiwse just sets the username.

Called by self$checkAuth and when settings are uploaded (in case the key/url weren't set)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RC$getUser(uname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uname}}{REDCap username}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{fullname} if success, \code{NULL} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RC-authfilter"></a>}}
\if{latex}{\out{\hypertarget{method-RC-authfilter}{}}}
\subsection{Method \code{authfilter()}}{
get auth filter exclusions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RC$authfilter()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a regex used to match auth filter exclusions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RC-getPID"></a>}}
\if{latex}{\out{\hypertarget{method-RC-getPID}{}}}
\subsection{Method \code{getPID()}}{
Lookup PIDs in REDCap
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RC$getPID(sampleid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sampleid}}{vector of length one or more of sample ids for which pids/mrns are needed}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
tibble of columns sample_id and pid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RC-updateCRF"></a>}}
\if{latex}{\out{\hypertarget{method-RC-updateCRF}{}}}
\subsection{Method \code{updateCRF()}}{
Upload CRFs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RC$updateCRF(sn, pool)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sn}}{A vector (length > 0) cartridge_sn in the sqlite database to be uploaded}

\item{\code{pool}}{\link{pool} scrapert database connection pool}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
character vector of cartridge_sn uploaded
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RC-CRFDelete"></a>}}
\if{latex}{\out{\hypertarget{method-RC-CRFDelete}{}}}
\subsection{Method \code{CRFDelete()}}{
REDCap Delete CRF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RC$CRFDelete(sn)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sn}}{the cartridge_sn in xpertdb}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} if successful
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RC-getPDF"></a>}}
\if{latex}{\out{\hypertarget{method-RC-getPDF}{}}}
\subsection{Method \code{getPDF()}}{
Get PDF from REDCap
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RC$getPDF(pid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pid}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
raw data (pdf)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RC-backup"></a>}}
\if{latex}{\out{\hypertarget{method-RC-backup}{}}}
\subsection{Method \code{backup()}}{
backup xpertdb and config.yml to REDCap file repository
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RC$backup()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
filename of .zip file in REDCap
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RC-restore"></a>}}
\if{latex}{\out{\hypertarget{method-RC-restore}{}}}
\subsection{Method \code{restore()}}{
restore backup from REDCap
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RC$restore()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
