% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCAuth.R
\name{RCAuth}
\alias{RCAuth}
\title{REDCap Authentication}
\usage{
RCAuth(config)
}
\arguments{
\item{config}{\link{Config} object}
}
\value{
a list of functions for plumber to reference
}
\description{
Performs authentication to REDCap
}
