% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCAuth.R
\name{RCAuthProcess}
\alias{RCAuthProcess}
\title{Process REDCap Responses}
\usage{
RCAuthProcess(req, res, config)
}
\arguments{
\item{req}{Plumber req object}

\item{res}{Plumber res object}

\item{config}{\link{Config} object}
}
\value{
\code{list(username = username)} if successful otherwise \code{list(err = "error message")} and/or sending
\code{res$status <- 401} (not authenticated). Will redirect if RCAuthAPI is posted as an arg after saving the config
}
\description{
Process the REDCap active link.
}
\details{
The active link should post it's response to \code{http://localhost:port/authep}.  No other variables are required.
This will also set RCApi if it's posted as an request arg
}
