% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCAuth.R
\name{RCAuthToken}
\alias{RCAuthToken}
\title{REDCap Token Authentication}
\usage{
RCAuthToken(token, config)
}
\arguments{
\item{token}{the REDCap Active link token}

\item{config}{\link{Config} object}
}
\value{
\code{list()} which if empty means we failed.  Otherwise returns the redcap username, project_id, etc.
}
\description{
REDCap Token Authentication
}
