% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCCRF.R
\name{RCCRF}
\alias{RCCRF}
\title{REDCap CRF Upload}
\usage{
RCCRF(config)
}
\arguments{
\item{config}{\link{Config} object}
}
\value{
a list of functions for plumber to reference
}
\description{
Performs CRF upload to REDCap
}
