% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCCRF.R
\name{RCCRFgetKey}
\alias{RCCRFgetKey}
\title{Get API Key}
\usage{
RCCRFgetKey(config)
}
\arguments{
\item{config}{\link{Config} Object}
}
\value{
API Key
}
\description{
Get the API Key, throw error is missing.  Check for URL as well.
}
