% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCCRF.R
\name{RCCRFgetRecordId}
\alias{RCCRFgetRecordId}
\title{REDCap Map row id to record id}
\usage{
RCCRFgetRecordId(df, config, key)
}
\arguments{
\item{df}{data frame with at least id, sample_id}

\item{config}{\link{Config} object}

\item{key}{REDCap CRF Api Key}
}
\value{
\code{df} with \code{record_id} column appended
}
\description{
REDCap Map row id to record id
}
