% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCPID.R
\name{RCPID}
\alias{RCPID}
\title{REDCap PID lookup}
\usage{
RCPID(config)
}
\arguments{
\item{config}{\link{Config} object}
}
\value{
a list of functions for plumber to reference
}
\description{
Performs pid lookup from REDCap
}
