% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCAuth.R
\name{RCsetUser}
\alias{RCsetUser}
\title{REDCap Util:  set user}
\usage{
RCsetUser(uname, config, req)
}
\arguments{
\item{uname}{REDCap username}

\item{config}{\link{Config} object}

\item{req}{Plumber \code{req} object}
}
\value{
\code{fullname} if success, \code{NULL} otherwise
}
\description{
If able to get full name from REDCap will set session cookie along with username, 
othersiwse just sets the username.
}
\details{
Called by \link{RCAuthProcess} and when settings are uploaded (in case the key/url were set)
}
