% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCCRF.R
\name{RCupdateCRF}
\alias{RCupdateCRF}
\title{Upload CRFs}
\usage{
RCupdateCRF(id, config)
}
\arguments{
\item{id}{A vector (length > 0) row ids in the sqlite database to be uploaded}

\item{config}{Scrapert config}
}
\value{
tibble 2 cols (id, uploaded) with the sqlite id, and an uploaded timestamp (UCT time POSIXct)
}
\description{
Upload CRFs
}
