% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scrapert.R
\docType{package}
\name{Scrapert}
\alias{Scrapert}
\title{Package Scrapert}
\description{
Scrapert is a R plumber based web app designed to pull information from Xpert pdf reports, serialize, and upload to a CRF
}
