% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCCRF.R
\name{RCCRFDelete}
\alias{RCCRFDelete}
\title{REDCap Delete CRF}
\usage{
RCCRFDelete(id, config)
}
\arguments{
\item{id}{the row id in xpertdb}

\item{config}{\link{Config} object}
}
\value{
\code{NULL} if successful
}
\description{
REDCap Delete CRF
}
